Move yaml file functions to separate file
