+++ tests/test_pipeline.py
@@ -3,7 +3,8 @@ import numpy as np
 import pytest
 from matchms import Pipeline
 from matchms.filtering import select_by_mz
-from matchms.Pipeline import create_workflow, load_workflow_from_yaml_file
+from matchms.Pipeline import create_workflow
+from matchms.yaml_file_functions import load_workflow_from_yaml_file
 from matchms.similarity import ModifiedCosine
